Scoping instances to containers
The annotation that scopes an instance to the application container is @Singleton.
This annotation will make the application container always provide the same instance regardless of whether the type is used as a dependency of another type or if it needs to be field injected.

Now, Hilt knows how to provide instances of LoggerLocalDataSource.
However, this time, the type has transitive dependencies! To provide an instance of LoggerLocalDataSource, Hilt also needs to know how to provide an instance of LogDao.

However, because LogDao is an interface, we cannot annotate its constructor with @Inject as interfaces don't have one. How can we tell Hilt how to provide instances of this type?