Hilt modules
Modules are used to add bindings to Hilt, or in other words, to tell Hilt how to provide instances of different types.
In Hilt modules, you include bindings for types that cannot be constructor injected such as interfaces or classes that are not contained in your project.

- Creating a Module
- Providing instances with @Provides

Let's continue refactoring the app to remove the ServiceLocator calls from the MainActivity.